cleanign thigns up a bit
